Move ContainerFormat enum to constants file
The ContainerFormat enumeration was moved from the dal/constants.ts file to a new constants.ts file in the components directory. This refactoring improves the organization and readability of the code by keeping related constants together in one place. In addition, some minor changes were made to code formatting in FileUploaderPage.tsx for clarity and consistency.